Move source tests, add tests for assemble
