+++ test/compile/data/datatestutil.ts
@@ -1,4 +1,6 @@
 import {DataComponent} from '../../../src/compile/data/data';
+import {Model} from '../../../src/compile/model';
+import {DataTable} from '../../../src/data';
 
 // FIXME refactor signature of assemble methods()
 
@@ -18,4 +20,15 @@ export function mockDataComponent(): DataComponent {
     stackScale: null,
     colorRank: null
   };
+}
+
+export function mockRenameModel(from: DataTable = null, to: string = null): Model {
+  return {
+    dataName: function(name) {
+      if (from && name === from) {
+        return to;
+      }
+      return name;
+    }
+  } as Model;
 }
